don't require 100GB for root filesystem
This changes the "LVM with cinder-volumes" layout such that we no longer
require 100GB for the root lv (via requiring 100GB for pv.01).  With
this change, we default both pv.01 and pv.02 to 1GB, and allow pv.01 to
grow up to 100GB and pv.02 to grow as large as possible.

Resolves: rhbz#1122753