BZ#1124545 - Write the values for skipping repo and subman
The interactive wizard gives the user the ability to skip either repo
setup or subscription manager setup. Neither of these choices are being
written to the answer file, which causes a stacktrace when run in non-
interactive mode. This patch writes those choices to the answer file,
and keeps them in subsequent runs.