BZ#1142182 - provisioned hosts have wrong timezone
https://bugzilla.redhat.com/show_bug.cgi?id=1142182

This adds support for setting timezone. The timezone will be set both
on the Foreman host and on the provisioned hosts. The timezone
defaults to the timezone currently set on the host where
staypuft-installer is run. The implementation should be forward
compatible for when Staypuft will run on EL 7 variants.