Set staypuft_ssh_public_key as a host parameter
and still keep it set in a smart class parameter in case some host group
includes the foreman::plugin::staypuft_client class.

The problem with previous approach was that if the host was
provisioned without foreman::plugin::staypuft_client puppet class
assigned to it, setting of the authorized key for puppetssh would fail
and the host would be "bricked" until either reprovisioned or set the
authorized key manually.