BZ#1180322 - errors on registration puppet run
https://bugzilla.redhat.com/show_bug.cgi?id=1180322

The main part of this issue is solved in Staypuft [1], but the first
puppet run could still fail due to pluginsync taking too long. The
pluginsync timeout is not configurable and is supposed to happen fast.
But we have a lot of modules and in virtualized testing environments the
puppetmaster machine serving the plugins and 4 or 5 clients fetching the
plugins are in fact served by the same hardware (usually just one HDD),
so this can take longer time and exceed the timeout. Only reasonable fix
seems to be just to run the installer twice to let the pluginsync
finish.

If the first pluginsync times out, the first report for that host will
contain errors, but the second run should finish the plugin sync and
therefore the host would be marked with "error" in Foreman for a very
short time, not for large part of OpenStack deployment duration, as it
was previously.

[1] https://github.com/theforeman/staypuft/pull/414